pyproject.toml
[project]
name = "agents-lucas-lindh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.7.4",
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "langchain==1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core==1.2.11",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph==1.0.8",
    "python-dotenv>=1.0.0",
]
